moved TryFrom implementation on HawkIdentifier to extractors.rs and removed use statements in middleware
